pyproject.toml
[project]
name = "agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "grandalf>=0.8",
    "langchain-community>=0.3.13",
    "langchain-openai>=0.2.14",
    "langgraph>=0.2.60",
    "langsmith>=0.2.6",
    "llama-cpp-python>=0.3.5",
    "python-dotenv>=1.0.1",
    "tavily-python>=0.5.0",
]
